requirements.txt
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
auto-py-to-exe==2.38.0
beautifulsoup4==4.12.3
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.3
cssselect2==0.8.0
cx-Freeze==6.15.7
cx-Logging==3.1.0
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
dlib==19.24.6
dnspython==2.7.0
Eel==0.16.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.13.1
Flask==3.0.2
fonttools==4.53.1
fpdf==1.7.2
frozendict==2.4.4
future==0.18.3
gevent==23.9.0.post1
gevent-websocket==0.10.1
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.5
lief==0.13.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
moviepy==1.0.3
multitasking==0.0.11
numpy==1.26.2
oscrypto==1.3.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
peewee==3.17.6
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.24.0
ply==3.11
proglog==0.1.10
proxy-tools==0.1.0
py2exe==0.13.0.0
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pydyf==0.11.0
pygame==2.5.2
Pygments==2.19.1
pyHanko==0.27.1
pyhanko-certvalidator==0.26.8
pyinstall==0.1.4
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pynput==1.7.6
pyparsing==3.1.1
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-nmap==0.7.1
pythonnet==3.0.4
pytz==2024.1
pywebview==5.3.2
pywin32-ctypes==0.2.2
PyYAML==6.0.2
qrcode==8.2
readchar==4.0.5
reportlab==4.2.0
requests==2.31.0
rich==13.9.4
rich-toolkit==0.12.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.4.4
starlette==0.41.3
statsmodels==0.14.2
svglib==1.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.0.4
uvicorn==0.34.0
virtualenv==20.25.1
watchfiles==1.0.3
webencodings==0.5.1
websockets==14.1
Werkzeug==3.0.1
whichcraft==0.6.1
xhtml2pdf==0.2.17
yfinance==0.2.43
zope.event==5.0
zope.interface==6.0
zopfli==0.2.3.post1
